pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-roster-v2-standalone"
version = "0.1.0"
description = "Standalone Universal Roster V2 with web chat workflow"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "openpyxl>=3.1",
  "jinja2>=3.1",
  "google-genai>=1.0.0",
  "google-auth>=2.20",
]

[project.optional-dependencies]
web = [
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "python-multipart>=0.0.9",
]
test = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
universal_roster_v2 = [
  "web/static/*.html",
  "web/static/*.js",
  "web/static/*.css",
  "web/static/legacy/*.html",
  "web/static/legacy/*.js",
  "web/static/legacy/*.css",
  "web/static/dist/*.html",
  "web/static/dist/assets/*",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
